+++ tools/prepare_vscode_tree.sh
@@ -7,7 +7,7 @@
 # After running this tool, you can compare ./ and ./tools/tmp/out/ to gather
 # new features.
 
-VSCODE_VERSION_GIT_TAG=1.66.2
+VSCODE_VERSION_GIT_TAG=1.67.0
 
 . $(dirname $0)/log.sh