+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [UNRELEASED]
+### Fixed
+- [DEPENDENCIES] Update for security fixes
+
+### Changed
+- [INTERNAL] Prepare update to vscode 1.67
+
+
 ## [1.3.21] - 20220918
 
 ### Fixed